SEACAS: Portabilty and static analysis fixes
(From Sierra)

+++ packages/seacas/applications/epu/epu.C
@@ -1694,7 +1694,7 @@ namespace {
     global_element_map.resize(tot_size);
 
     {
-      size_t error = 0;
+      int error = 0;
       size_t offset = 0;
       for (int p = 0; p < part_count; p++) {
 	ExodusFile id(p);
@@ -1875,7 +1875,7 @@ namespace {
     global_node_map.resize(tot_size);
 
     size_t offset = 0;
-    size_t error = 0;
+    int error = 0;
     for (int p = 0; p < part_count; p++) {
       ExodusFile id(p);
       error = ex_get_node_num_map(id, TOPTR(global_node_numbers[p]));
